+++ test/jest/__mock__/stripesInventoryComponents.mock.js
@@ -1,0 +1,3 @@
+jest.mock('@folio/stripes-inventory-components', () => ({
+  ...jest.requireActual('@folio/stripes-inventory-components'),
+}));