+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [UIQM-647](https://issues.folio.org/browse/UIQM-647) Import `useUserTenantPermissions` from `@folio/stripes/core`.
 * [UIQM-563](https://issues.folio.org/browse/UIQM-563) Allow a user to move 00X fields.
+* [UIQM-563](https://issues.folio.org/browse/UIQM-655) Add a `required` property to `Select` type fields for screen readers.
 
 ## [8.0.1] (https://github.com/folio-org/ui-quick-marc/tree/v8.0.1) (2024-04-18)